Fix #17
This is very much a "work for me" kind of fix. It's up to you,
maintainer, but it probably serves more as a reference than a fix that
should be merged as-is. I'm neither a Java guy nor have I taken the time
to understand the architecture of this plugin in-depth, so I'm not sure
how to implement it in a "clean" way.